requirements.txt
# Web app and mqtt client
asgiref==3.8.1
Django==5.1.6
djangorestframework==3.15.2
psycopg2==2.9.10
sqlparse==0.5.3
channels==4.2.0
paho-mqtt==2.1.0
channels-redis==3.3.1
